+++ src/configs/json.ts
@@ -1,19 +1,26 @@
 import jsoncPlugin from 'eslint-plugin-jsonc';
 import * as jsoncParser from 'jsonc-eslint-parser';
 
-import type { Config, Rules } from '../core';
+import {
+  type Config,
+  type Rules,
 
-import { withConfig } from '../core';
+  GLOB_JSON,
+  GLOB_JSONC,
+} from '../core';
 
 const jsoncRecommendedRules = jsoncPlugin.configs['recommended-with-json'].rules as Rules;
 
-const JSONC_ALLOW_COMMENTS_FILES = [
+const GLOB_PACKAGE_JSON = '**/package.json';
+
+const JSONC_FILES = [
+  GLOB_JSONC,
   '**/.vscode/*.json',
   '**/tsconfig.json',
   '**/tsconfig.*.json',
 ];
 
-export const poupeJsonRules: Rules = {
+const poupeJsonRules: Rules = {
   // Consistent formatting
   'jsonc/indent': ['error', 2],
   'jsonc/comma-dangle': ['error', 'never'],
@@ -28,7 +35,12 @@ export const poupeJsonRules: Rules = {
   'jsonc/no-comments': 'error',
 };
 
-export const poupePackageJsonRules: Rules = {
+const poupeJsoncRules: Rules = {
+  ...poupeJsonRules,
+  'jsonc/no-comments': 'off',
+};
+
+const poupePackageJsonRules: Rules = {
   ...poupeJsonRules,
   // Sort package.json keys in standard order
   'jsonc/sort-keys': [
@@ -76,11 +88,11 @@ export const poupePackageJsonRules: Rules = {
   ],
 };
 
-export const jsoncRecommended: Config[] = withConfig(
+export const poupeJsonConfigs: Config[] = [
   {
     name: 'poupe/json',
-    files: ['**/*.json'],
-    ignores: ['**/package.json'],
+    files: [GLOB_JSON],
+    ignores: [GLOB_PACKAGE_JSON],
     plugins: {
       jsonc: jsoncPlugin,
     },
@@ -94,7 +106,7 @@ export const jsoncRecommended: Config[] = withConfig(
   },
   {
     name: 'poupe/package-json',
-    files: ['**/package.json'],
+    files: [GLOB_PACKAGE_JSON],
     plugins: {
       jsonc: jsoncPlugin,
     },
@@ -107,10 +119,11 @@ export const jsoncRecommended: Config[] = withConfig(
     },
   },
   {
-    name: 'poupe/allow-json-comments',
-    files: JSONC_ALLOW_COMMENTS_FILES,
+    name: 'poupe/jsonc',
+    files: JSONC_FILES,
     rules: {
-      'jsonc/no-comments': 'off',
+      ...jsoncRecommendedRules,
+      ...poupeJsoncRules,
     },
   },
-);
+];