+++ src/configs/perfectionist.ts
@@ -3,11 +3,12 @@ import perfectionist from 'eslint-plugin-perfectionist';
 import {
   type Config,
   type Rules,
-  withConfig,
+
+  GLOB_SRC,
+  GLOB_VUE,
 } from '../core';
 
-// Perfectionist rules for import/export sorting
-export const poupePerfectionistRules: Rules = {
+const poupePerfectionistRules: Rules = {
   'perfectionist/sort-imports': [
     'error',
     {
@@ -15,8 +16,9 @@ export const poupePerfectionistRules: Rules = {
       order: 'asc',
       ignoreCase: true,
       internalPattern: ['^~/.+', '^@/.+', String.raw`^\.\./.+`, String.raw`^\./.+`],
-      newlinesBetween: 'always',
+      newlinesBetween: 'ignore',
       maxLineLength: undefined,
+      partitionByNewLine: true,
       groups: [
         'type',
         ['builtin', 'external'],
@@ -57,6 +59,7 @@ export const poupePerfectionistRules: Rules = {
       ignoreAlias: false,
       specialCharacters: 'keep',
       groupKind: 'mixed',
+      partitionByNewLine: true,
     },
   ],
   'perfectionist/sort-named-exports': [
@@ -81,11 +84,11 @@ export const poupePerfectionistRules: Rules = {
   ],
 };
 
-export const perfectionistRecommended: Config[] = withConfig({
+export const poupePerfectionistConfig: Config = {
   name: 'poupe/perfectionist',
-  files: ['**/*.{js,mjs,cjs,ts,mts,cts,jsx,tsx,vue}'],
+  files: [GLOB_SRC, GLOB_VUE],
   plugins: {
     perfectionist,
   },
   rules: poupePerfectionistRules,
-});
+};