+++ src/core/index.ts
@@ -1,2 +1,3 @@
 export * from './config';
+export * from './globs';
 export * from './utils';